Refactor ability system and add UI for ability slots
Refactored the ability system by introducing new effect and targeting strategy factories, replacing and reorganizing several scripts. Added UI components and scene objects for ability slots, including icon support and editor improvements. Updated TestAbilityData and DemoScene to use the new structure, and removed obsolete scripts.

+++ Assets/AbilitySystem/Scripts/Damage/Enemy.cs
@@ -3,11 +3,11 @@ using UnityEngine;
 
 public class Enemy : MonoBehaviour, IDamageable
 {
-    public int Health = 50;
+    public float Health = 50;
 
     readonly List<IEffect<IDamageable>> _activeEffects = new List<IEffect<IDamageable>>();
 
-    public void TakeDamage(int damage)
+    public void TakeDamage(float damage)
     {
         Health -= damage;
         if (Health <= 0)
